requirements.txt
streamlit
google-generativeai
nltk
python-dotenv
plotly
pandas